+++ packages/taro-cli/src/h5/index.ts
@@ -425,7 +425,7 @@ class Compiler {
             )
 
             node.body = toAst(buildFuncBody(pages), { preserveComments: true })
-
+          } else {
             node.body.body = compact([
               hasComponentDidHide && isComponentWillUnmount && callComponentDidHideNode,
               ...node.body.body,
@@ -803,7 +803,7 @@ class Compiler {
                 )
 
                 node.body = toAst(buildFuncBody(pages), { preserveComments: true })
-
+              } else {
                 node.body.body = compact([
                   hasComponentDidHide && isComponentWillUnmount && callComponentDidHideNode,
                   ...node.body.body,