+++ packages/taro-components/src/components/image/index.js
@@ -62,6 +62,7 @@ class Image extends Nerv.Component {
       mode,
       onError,
       lazyLoad,
+      imgProps,
       ...reset
     } = this.props
     const { aspectFillMode } = this.state
@@ -97,6 +98,7 @@ class Image extends Nerv.Component {
             src={src}
             onLoad={this.imageOnLoad}
             onError={onError}
+            {...imgProps}
           />
         )}
       </div>